Added comments and tidied everything up a little. A lot more to clean up before adding more
