+++ .gitignore
@@ -6,3 +6,4 @@
 erl_crash.dump
 /config/*.secret.exs
 /.elixir_ls
+.DS_Store
