+++ Makefile
@@ -1,8 +1,9 @@
+serve:
+	iex -S mix phx.server
+
 start_psql:
 	pg_ctl -D /usr/local/var/postgres start
 
-serve:
-	mix phx.server
 
 phxr:
 	mix phx.routes
