--- lib/hello/topic/comments.ex
@@ -1,117 +1,0 @@
-defmodule Hello.Topic.Comments do
-  @moduledoc """
-  The Topic.Comments context.
-  """
-
-  import Ecto.Query, warn: false
-  alias Hello.Repo
-  alias HelloWeb.Topic
-  alias Hello.Topic.Comments.Topic.Comment
-
-  @doc """
-  Returns the list of topic_comments.
-
-  ## Examples
-
-      iex> list_topic_comments()
-      [%Comment{}, ...]
-
-  """
-  def list_topic_comments do
-    Repo.all(Comment)
-  end
-
-  def list_topic_comments(topic_id) do
-    (from c in Comment,
-      where: c.topic_id == ^topic_id
-    )
-    |> Repo.all()
-
-    # Comment |> where([comment], comment.topic_id in 3) |> Repo.all()
-
-    # Repo.all(Comment, topic_id: topic_id)
-  end
-
-  @doc """
-  Gets a single comment.
-
-  Raises `Ecto.NoResultsError` if the Comment does not exist.
-
-  ## Examples
-
-      iex> get_comment!(123)
-      %Comment{}
-
-      iex> get_comment!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_comment!(id), do: Repo.get!(Comment, id)
-
-  def get_topic!(topic_id), do: Repo.get!(Topic, topic_id)
-
-  @doc """
-  Creates a comment.
-
-  ## Examples
-
-      iex> create_comment(%{field: value})
-      {:ok, %Comment{}}
-
-      iex> create_comment(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_comment(attrs \\ %{}) do
-    %Comment{}
-    |> Comment.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a comment.
-
-  ## Examples
-
-      iex> update_comment(comment, %{field: new_value})
-      {:ok, %Comment{}}
-
-      iex> update_comment(comment, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_comment(%Comment{} = comment, attrs) do
-    comment
-    |> Comment.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a Comment.
-
-  ## Examples
-
-      iex> delete_comment(comment)
-      {:ok, %Comment{}}
-
-      iex> delete_comment(comment)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_comment(%Comment{} = comment) do
-    Repo.delete(comment)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking comment changes.
-
-  ## Examples
-
-      iex> change_comment(comment)
-      %Ecto.Changeset{source: %Comment{}}
-
-  """
-  def change_comment(%Comment{} = comment) do
-    Comment.changeset(comment, %{})
-  end
-end
