+++ lib/hello_web/templates/comment/new.html.eex
@@ -1,3 +1,1 @@
 <h1>New Comment</h1>
-
-<%= render "form.html", Map.put(assigns, :action, Routes.topic_comment_path(@conn, :create)) %>
