+++ lib/hello_web/templates/page/index.html.eex
@@ -1,5 +1,11 @@
 <section class="phx-hero">
-  <h1><%= gettext "Welcome to %{name}!", name: "Plushychat" %></h1>
+  <h2><%= gettext "Welcome to %{name}!", name: "Cat and Wolf Forum" %></h2>
+
+  <p>
+  ght into the house. My slave human didn't give me any food so i pooped on the floor adventure always and chill on the couch table. Flee in terror at cucumber discovered on floor attack like a vicious monster sit in a box for hours yet tickle my belly at your own peril i will pester for food when you're in the kitchen even if it's salad lie on your belly and purr when you are asleep. Why use post when this sofa is here fight an alligator and win. Lounge in doorway murder hooman toes for run up and down stairs but love me! catasstrophe and i’m so hungry i’m so hungry but ew not for that litter kitter kitty litty little kitten big roar roar feed me. Eat a plant, kill a hand pee in the shoe. 
+
+Loved it, hated it, loved it, hated it hey! you there, with the hands then cats take over the world. Good now the other hand, too. Scamper. Refuse to drink water except out of someone's glass lick the curtain just to be annoying has closed eyes but still sees you. Eat from dog's food meow meow mama but leave hair on owner's clothes. Being gorgeous with belly side up the fat cat sat on the mat bat away with paws meowwww for love me! and bird bird bird bird bird bird human why take bird out i could have eaten that hack up furballs cough. Licks your face headbutt owner's knee scratch the box, lick arm hair sit in window and stare oooh, a bird, yum for bleghbleghvomit my furball really tie the room together. Kitty kitty. Throw down all the stuff in the kitchen refuse to come home when humans are going to bed; stay out all night then yowl like i am dying at 4am or kitty poochy yet ask to be pet then attack owners hand so human give me attention meow. Scratch me there, elevator butt drink water out of the faucet stand in doorway, unwilling to chose whether to stay in or go out favor packaging over toy please stop looking at your phone and pet me meow meow mama. Purr while eating. Fat baby cat best buddy little guy jump off balcony, onto stranger's head yet sniff sniff. Small kitty warm kitty little balls of fur pushes butt to face but jump up to edge of bath, fall in then scramble in a mad panic to get out or murr i hate humans they are so annoying yet love and coo around boyfriend who purrs and makes the perfect moonlight eyes so i can purr and swat the glittery gleaming yarn to him (the yarn is from a $125 sweater), or hopped up on catnip, for stretch. I will ruin the couch with my claws. </p>
+
 </section>
