+++ lib/hello_web/templates/topic/comment_form.html.eex
@@ -5,25 +5,13 @@
     </div>
   <% end %>
 
-  <%= label f, :user_id %>
-  <%= number_input f, :user_id %>
-  <%= error_tag f, :user_id %>
 
-  <%= label f, :topic_id %>
-  <%= number_input f, :topic_id %>
-  <%= error_tag f, :topic_id %>
 
   <%= label f, :comment %>
   <%= text_input f, :comment %>
   <%= error_tag f, :comment %>
 
-  <%= label f, :likes %>
-  <%= number_input f, :likes %>
-  <%= error_tag f, :likes %>
 
-  <%= label f, :reply_to %>
-  <%= number_input f, :reply_to %>
-  <%= error_tag f, :reply_to %>
 
   <div>
     <%= submit "Save" %>
