+++ lib/hello_web/templates/topic/show.html.eex
@@ -52,8 +52,6 @@
 
 
 
-
-
 <div>
   <h1>New Comment</h1>
