+++ lib/hello_web/templates/user/index.html.eex
@@ -1,14 +1,6 @@
-<h1>Listing Users</h1>
+<h2>Users</h2>
 
-<table>
-  <thead>
-    <tr>
-      <th>Username</th>
-      <th>Email</th>
 
-      <th></th>
-    </tr>
-  </thead>
    <div class="row">
   <%= for user <- @users do %>
 
@@ -34,20 +26,7 @@
 <% end %>
    </div>
           
-  <tbody>
-<%= for user <- @users do %>
 
 
 
-    <tr>
-      <td><%= user.username %></td>
-      <td><%= user.email %></td>
-
-      <td>
-        <%= link "Show", to: Routes.user_path(@conn, :show, user) %>
-      </td>
-    </tr>
-<% end %>
-  </tbody>
-</table>
